general game logic set, turns, and game components
Game window displays score cards for each logged in user. Will display in color blue the current players card when its their turn.

Users can already write a word in the imput component and the game component will call the dictionary API to check if the word exists in the english dictionary. Points are awarded if it exists and the it will be the next player's turn. If the users input does not exists in the english dictionary an alert will let him know.

Per letter point systemis still not implemented